src/main/java/org/example/lms_project/Model/Borrow.java
package org.example.lms_project.Model;

import java.util.Date;

public class Borrow {
    private Long id;
    private Long userId;
    private Long bookId;
    private Date borrowDate;
    private Date returnDate;
    private String status;

    public Borrow() {}

    public Borrow(Long id, Long userId, Long bookId, Date borrowDate, Date returnDate, String status) {
        this.id = id;
        this.userId = userId;
        this.bookId = bookId;
        this.borrowDate = borrowDate;
        this.returnDate = returnDate;
        this.status = status;
    }
    public Borrow(Long userId, Long bookId, Date returnDate, String status) {
        this.userId = userId;
        this.bookId = bookId;
        this.borrowDate = new Date();
        this.returnDate = returnDate;
        this.status = status;
    }

    // Getters and Setters
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getUserId() {
        return userId;
    }

    public void setMemberId(Long userId) {
        this.userId = userId;
    }

    public Long getBookId() {
        return bookId;
    }

    public void setBookId(Long bookId) {
        this.bookId = bookId;
    }

    public Date getBorrowDate() {
        return borrowDate;
    }

    public void setBorrowDate(Date borrowDate) {
        this.borrowDate = borrowDate;
    }

    public Date getReturnDate() {
        return returnDate;
    }

    public void setReturnDate(Date returnDate) {
        this.returnDate = returnDate;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }
}

